Reduce size of built Docker image
Shrinks the swiftbox image from 1.22 GB to 849.3 MB.

The result of each Dockerfile line is saved to its own fs layer, so temporary files take up space in the image if they aren't removed in the same line they are created in. Also apt-get clean frees up a little space.