Prevent SIGPIPE crashes
When `send`ing on a socket which has been closed on the remote end, the
process is sent SIGPIPE by default, which causes the process to be
terminated when it is unhandled. This change modifies the behavior so
that the `send` call will return an error instead of the signal being
sent.

On Mac, this is done by setting a socket option, while on Linux it is
done by passing a flag to `send`.